requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
black==23.12.0
click==8.1.7
Django==5.0
django-dotenv==1.4.2
django-filter==23.5
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.6
isort==5.13.1
Markdown==3.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
phonenumbers==8.13.26
platformdirs==4.1.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
yarl==1.9.4
